Fixed authentication and middleware

+++ app/Http/Controllers/ListingController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Listing;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\Rule;
 
 class ListingController extends Controller
@@ -47,6 +49,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -55,12 +59,22 @@ class ListingController extends Controller
     // Edit listing data
     public function edit(Listing $listing)
     {
+        // Guard clause to make sure logged in user is owner
+        if ($listing->user_id !== auth()->id()) {
+            abort('403', 'Unauthorized action');
+        }
+
         return view('listings.edit', ['listing' => $listing]);
     }
 
     // Update listing
     public function update(Request $request, Listing $listing)
     {
+        // Guard clause to make sure logged in user is owner
+        if ($listing->user_id !== auth()->id()) {
+            abort('403', 'Unauthorized action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -83,8 +97,20 @@ class ListingController extends Controller
     // Delete listing
     public function destroy(Listing $listing)
     {
+        // Guard clause to make sure logged in user is owner
+        if ($listing->user_id !== auth()->id()) {
+            abort('403', 'Unauthorized action');
+        }
+        
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully!');
     }
+
+    // Manage listing
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => User::find(auth()->id())->listings()->get()]);
+    }
+
 }
